系统维护-基本--

+++ src/components/xtwh/ggdmwh/xqdmwh/mainTable.vue
@@ -3,8 +3,9 @@
         stripe
         border
         highlight-current-row
+        v-loading="loading"
         @row-click="handleCurrentChange"
-        height="350px"
+        :height="tableHeight"
         :data="tableData"
         style="width: 100%"
         :default-sort = "{prop: 'xqdm', order: 'descending'}">
@@ -53,6 +54,8 @@ export default {
     data() {
         return {
             tableData:[],
+            loading: true,
+            tableHeight: window.innerHeight * 0.75 ,
         }
     },
     methods: {
@@ -70,18 +73,30 @@ export default {
             ////通过改变每次的参数解决路由跳转失效的问题
       },
       handleDelete(index, row) {
-          this.deleteData(row.xqdm);
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                this.deleteData(row.xqdm);
+            }).catch(() => {
+                this.$message({
+                    type: 'info',
+                    message: '已取消删除'
+                });          
+            });
       },
       deleteData(xqdm){
             var _this=this;
             //需要处理异步请求的问题
             this.axios.get('jwc/SysXq/delete?xqdm='+xqdm)
                 .then(function (response) {
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    //refresh
+                    _this.$message({ message: '删除校区代码成功',showClose: true,type: 'success' });
                 })
                 .catch(function (error) {
                     console.log(error);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息失败", type:"error"})
+                    _this.$message({ message: '删除校区代码失败，请检查网络',showClose: true,type: 'error' });
                 });
       },
       //将数据库存储的状态数值，格式化为汉字
@@ -91,6 +106,7 @@ export default {
       },
       getData(){
             var _this=this;
+            _this.loading = true;
             //需要处理异步请求的问题
             this.axios.get('jwc/SysXq/getAll')
                 .then(function (response) {
@@ -98,11 +114,13 @@ export default {
                     //将获取到的数据以数组形式传递出去
                     var dataList=response.data;
                     _this.tableData=dataList;
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    _this.$notify({title:"获取校区信息", message:"获取校区信息成功", type:"success"});
+                    _this.loading = false;
                 })
                 .catch(function (error) {
                     console.log(error);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    _this.$notify({title:"获取校区信息", message:"获取校区信息失败", type:"error"});
+                    _this.loading = false;
                 });
       }
     }
