+++ src/components/xtwh/ggdmwh/xydmwh/mainTable.vue
@@ -4,7 +4,8 @@
         border
         highlight-current-row
         @row-click="handleCurrentChange"
-        height="350px"
+        :height="tableHeight"
+        v-loading="loading"
         :data="tableData"
         style="width: 100%"
         :default-sort = "{prop: 'xqdm', order: 'descending'}"
@@ -58,7 +59,9 @@ export default {
     },
     data() {
       return {
-         tableData: []
+          tableHeight: window.innerHeight * 0.75 ,
+          loading: true,
+          tableData: []
       }
     },
     methods: {
@@ -80,7 +83,18 @@ export default {
             params:{ val:row ,change_id: row.xydm+new Date().getSeconds(), type: 'change'}});
       },
       handleDelete(index, row) {
-        this.deleteData(row.xydm);
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                this.deleteData(row.xydm);
+            }).catch(() => {
+                this.$message({
+                    type: 'info',
+                    message: '已取消删除'
+                });          
+            });
       },
       deleteData(xydm){
             var _this=this;
@@ -91,15 +105,16 @@ export default {
                     //将获取到的数据以数组形式传递出去
                     alert(response.data);
                     _this.$router.go(0);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    _this.$message({ message: '成功删除学院信息', type: 'success' });
                 })
                 .catch(function (error) {
                     console.log(error);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    _this.$message({ message: '删除学院信息失败', type: 'error' });
                 });
       },
       getData(){
         var _this=this;
+        _this.loading = true;
         //需要处理异步请求的问题
         this.axios.get('jwc/SysXy/getAllList')
             .then(function (response) {
@@ -107,11 +122,13 @@ export default {
                 //将获取到的数据以数组形式传递出去
                 var dataList=response.data;
                 _this.tableData=dataList;
-                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.$notify({title:"获取学院信息", message:"获取学院信息成功", type:"success"})
+                _this.loading = false;
             })
             .catch(function (error) {
                 console.log(error);
-                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.$notify({title:"获取学院信息", message:"获取学院信息失败", type:"error"});
+                _this.loading = false;
             });
       }
     }
