+++ src/components/xtwh/ggdmwh/zydmwh/mainTable.vue
@@ -2,8 +2,9 @@
     <el-table
         stripe
         border
+        v-loading="loading"
         highlight-current-row
-        height="500px"
+        :height="tableHeight"
         @row-click="handleCurrentChange"
         :data="tableData"
         :default-sort = "{prop: 'xqdm', order: 'descending'}"
@@ -109,7 +110,9 @@ export default {
     },
     data() {
       return {
-         tableData: []
+         tableData: [],
+         tableHeight: window.innerHeight * 0.75 ,
+         loading: true,
       }
     },
     methods: {
@@ -131,9 +134,20 @@ export default {
       //  this.$router.replace({name: 'zydmRightForm',params:{ val:row ,change_id: row.zydm, type: 'change'}});
       //},
       handleDelete(index, row) {
-          this.deleteData(row.zydm);
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                this.deleteData(row.zydm);
+            }).catch(() => {
+                this.$message({
+                    type: 'info',
+                    message: '已取消删除'
+                });          
+            });
       },
-    deleteData(zydm){
+      deleteData(zydm){
         var _this=this;
         //需要处理异步请求的问题
         this.axios.get('jwc/SysZy/delete?zydm='+ zydm)
@@ -141,18 +155,19 @@ export default {
                 //将response获得的数据进行处理
                 //将获取到的数据以数组形式传递出去
                 alert(response.data);
-                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.$message({ message: '成功删除专业代码', type: 'success' });
             })
             .catch(function (error) {
                 console.log(error);
-                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.$message({ message: '删除专业代码失败', type: 'error' });
             });
-    },
+      },
       getData(xydm,zy){
             if(!xydm){
                 xydm=0;
             }
             var _this=this;
+            _this.loading = true;
             //需要处理异步请求的问题
             this.axios.get('jwc/SysZy/WhSearch', {//通过这种方式解决模糊匹配后台报空指针异常的问题
                 params: {
@@ -160,17 +175,19 @@ export default {
                     zymc:zy,
                 }
             })
-                .then(function (response) {
-                    //将response获得的数据进行处理
-                    //将获取到的数据以数组形式传递出去
-                    var dataList=response.data;
-                    _this.tableData=dataList;
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
-                })
-                .catch(function (error) {
-                    console.log(error);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
-                });
+            .then(function (response) {
+                //将response获得的数据进行处理
+                //将获取到的数据以数组形式传递出去
+                var dataList=response.data;
+                _this.tableData=dataList;
+                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.loading = false;
+            })
+            .catch(function (error) {
+                console.log(error);
+                _this.$notify({title:"获取专业信息", message:"获取专业信息失败", type:"error"})
+                _this.loading = false;
+            });
       }
     }
 }
