+++ src/components/xtwh/jhzmkdmwh/khfsdmwh/mainTable.vue
@@ -3,8 +3,9 @@
         stripe=true
         border=true
         highlight-current-row
+        v-loading="loading"
         @row-click="handleCurrentChange"
-        height="350px"
+        :height="tableHeight"
         :data="tableData"
         style="width: 100%"
         :default-sort = "{prop: 'xqdm', order: 'descending'}"
@@ -53,7 +54,9 @@ export default {
     },
     data() {
       return {
-         tableData: []
+         tableData: [],
+         tableHeight: window.innerHeight * 0.75 ,
+         loading: true,
       }
     },
     methods: {
@@ -75,7 +78,19 @@ export default {
             params:{ val:row ,change_id: row.khfsdm+new Date().getSeconds(), type: 'change'}});
       },
       handleDelete(index, row) {
-        this.deleteData(row.khfsdm);
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                this.deleteData(row.khfsdm);
+            }).catch(() => {
+                this.$message({
+                    type: 'info',
+                    message: '已取消删除'
+                });          
+            });
+        
       },
       deleteData(dm){
             var _this=this;
@@ -86,15 +101,16 @@ export default {
                     //将获取到的数据以数组形式传递出去
                     alert(response.data);
                     //_this.$router.go(0);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    _this.$message({ message: '成功删除考核方式', type: 'success' });
                 })
                 .catch(function (error) {
                     console.log(error);
-                    _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                    _this.$message({ message: '删除考核方式失败', type: 'error' });
                 });
       }, 
       getData(){
         var _this=this;
+        _this.loading = true;
         //需要处理异步请求的问题
         this.axios.get('jwc/DmKhfs/getAll')
             .then(function (response) {
@@ -102,11 +118,13 @@ export default {
                 //将获取到的数据以数组形式传递出去
                 var dataList=response.data;
                 _this.tableData=dataList;
-                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.$notify({title:"获取考核方式", message:"获取考核方式成功", type:"success"});
+                _this.loading = false;
             })
             .catch(function (error) {
                 console.log(error);
-                _this.$notify({title:"获取专业信息", message:"获取专业信息成功", type:"success"})
+                _this.$notify({title:"获取考核方式", message:"获取考核方式失败", type:"error"});
+                _this.loading = false;
             });
       }
     }
